requirements.txt
django
djangorestframework
environ
django-cors-headers